New implementation - try

+++ src/include/types.h
@@ -62,6 +62,6 @@ namespace aos {
 		~File();
 
 		size_t getSize() const { return _size; }
-		bool read(const uint16_t, const uint16_t, const uint8_t*);
+		bool read(const size_t, const size_t, const uint8_t*);
 	};
 };
